+++ lib/seq_as_enum.rb
@@ -5,6 +5,8 @@ require_relative "seq_as_enum/version"
 module SeqAsEnum
   DEFAULT_SEP = '_'.freeze
 
+  private
+
   def seq_as_enum(*names, init: 0, data_as: nil, prefix: nil, sep: DEFAULT_SEP)
     raise ArgumentError, 'Cannot use prefix option with data_as option' if prefix && data_as
     raise ArgumentError, 'Cannot use sep option without prefix option' if sep != DEFAULT_SEP && !prefix
@@ -15,22 +17,15 @@ module SeqAsEnum
     if data_as
       class_name = "#{data_as}Data"
       data_class = const_defined?(:Data) ? Data.define(*names) : Struct.new(*names, keyword_init: true)
-      const_set!(class_name, data_class)
+      const_set(class_name, data_class)
       data = data_class.new(**key_values)
-      const_set!(data_as, data)
+      const_set(data_as, data)
     else
       key_values.each do |name, value|
         sep = '' unless sep
         name = "#{prefix}#{sep}#{name}".to_sym if prefix
-        const_set!(name, value)
+        const_set(name, value)
       end
     end
-  end
-
-  private
-
-  def const_set!(sym, obj)
-    raise "already initialized constant: #{self.name}::#{sym}" if const_defined?(sym)
-    const_set(sym, obj)
   end
 end