+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Allow override constant
+
 ## [0.2.0] - 2023-06-06
 
 - Use Struct instead of Data for Ruby 3.1.x and 3.0.x environments.